Refactor, stubbing out Splash screen
