+++ clients/cli/src/ui/splash.rs
@@ -1,0 +1,14 @@
+//! Splash screen rendering module.
+
+use ratatui::Frame;
+use ratatui::layout::Alignment;
+use ratatui::style::{Color, Modifier, Style};
+use ratatui::widgets::{Block, Borders, Paragraph};
+
+pub fn render_splash(frame: &mut Frame) {
+    let text = Paragraph::new("NEXUS SPLASH SCREEN")
+        .alignment(Alignment::Center)
+        .style(Style::default().fg(Color::Cyan).add_modifier(Modifier::BOLD))
+        .block(Block::default().borders(Borders::ALL));
+    frame.render_widget(text, frame.size());
+}
